@@ -7,7 +7,8 @@
     "license": "MIT",
 
     "require": {
-        "php": ">=7.3"
+        "php": ">=7.4",
+        "phpseclib/phpseclib": "~3.0"
     },
 
     "require-dev": {
