@@ -23,6 +23,16 @@ namespace Ubiq;
  */
 class KeyManager
 {
+    /**
+     * Gets the configured TTL (or default) from configuration
+     *
+     * @param Credentials $creds   Credentials object to operate on
+     * 
+     * @return int TTL seconds
+     */
+    public static function getCacheTTL(Credentials $creds) {
+        return $creds::$config['key_caching']['ttl_seconds'] ?? 1800;
+    }
 
     /**
      * Gets the default key for a dataset
@@ -35,14 +45,19 @@ class KeyManager
      */
     public function getKeyDefault(Credentials $creds, Dataset $dataset)
     {
-        $cache = $creds::$cachemanager->get(
+        $cache = $creds::$cachemanager::get(
             CacheManager::CACHE_TYPE_KEYS, $dataset->name . '-keys-default'
         );
-        
-        if (!empty($cache)) {
-            ubiq_debug($creds, 'Found default key of ' . $cache['key_idx'] . ' for ' . $dataset->name);
 
-            return $this->getKey($creds, $dataset, $cache['_key_enc']);
+        if (!empty($cache)) {
+            ubiq_debug($creds, 'Found default key cached of ' . $cache['key_idx'] . ' for ' . $dataset->name);
+
+            if ($dataset->type == DatasetManager::DATASET_TYPE_UNSTRUCTURED) {
+                return $this->getKey($creds, $dataset, $cache['_key_enc']);
+            }
+            elseif ($dataset->type == DatasetManager::DATASET_TYPE_STRUCTURED) {
+                return $this->getKey($creds, $dataset, $cache['_key_enc']);
+            }
         }
 
         return false;        
@@ -67,18 +82,18 @@ class KeyManager
 
         ubiq_debug($creds, 'Looking for cached key for ' . $dataset->name . ' for key ' . $key_idx);
 
-        $cache = $creds::$cachemanager->get(
+        $cache = $creds::$cachemanager::get(
             CacheManager::CACHE_TYPE_KEYS,
             $dataset->name . '-keys-' . $key_idx
         );
         
         if (!empty($cache)) {
-            ubiq_debug($creds, 'Found cached key for ' . $dataset->name . ' for key ' . $key_idx);
+            ubiq_debug($creds, 'Found key cached for ' . $dataset->name . ' for key ' . $key_idx);
                 
             $return = $cache;
 
             // if caching and encrypting, decrypt the key
-            if ($creds->config['key_caching']['encrypt']) {
+            if ($creds::$config['key_caching']['encrypt']) {
                 ubiq_debug($creds, 'Cached keys are encrypted; decrypting prior to returning');
                 
                 $pkey = openssl_pkey_get_private(
@@ -124,7 +139,7 @@ class KeyManager
     ) {
         $key_idx = $cache_data['key_idx'];
 
-        ubiq_debug($creds, 'Force disable cache ' . ($no_cache ? 'true' : 'false'));
+        ubiq_debug($creds, 'Force disable key cache ' . ($no_cache ? 'true' : 'false'));
         ubiq_debug($creds, 'Managing key for ' . $dataset->name . ' for key ' . $key_idx);
 
         // if the key was bad, return an exception
@@ -154,21 +169,29 @@ class KeyManager
 
         if ($this->_shouldCache($creds, $dataset) && !$no_cache) {
             ubiq_debug($creds, 'Caching key for ' . $dataset->name . ' for key ' . $key_idx);
+            ubiq_debug($creds, 'Setting TTL for ' . $dataset->name . ' for key ' . $key_idx . ' to ' . self::getCacheTTL($creds) . ' seconds');
 
-            if (!$creds->config['key_caching']['encrypt']) {
+            if (!$creds::$config['key_caching']['encrypt']) {
                 ubiq_debug($creds, 'Cached keys are NOT encrypted; decrypting prior to cache');
             
                 $cache_data['_key_raw'] = $key_raw;
             }
 
-            $creds::$cachemanager->set(
+            $creds::$cachemanager::set(
                 CacheManager::CACHE_TYPE_KEYS,
-                $dataset->name . '-keys-' . $key_idx, $cache_data
+                $dataset->name . '-keys-' . $key_idx,
+                $cache_data
+            );
+
+            $creds::$cachemanager::setTTL(
+                CacheManager::CACHE_TYPE_KEYS,
+                $dataset->name . '-keys-' . $key_idx,
+                strtotime('+' . self::getCacheTTL($creds) . ' seconds')
             );
         }
 
         $cache_data['_key_raw'] = $key_raw;
-        // print_r($creds::$cachemanager::$caches);
+
         return $cache_data;
     }
 
@@ -182,12 +205,16 @@ class KeyManager
      */
     private function _shouldCache(Credentials $creds, Dataset $dataset)
     {
-        $type = $dataset->type ?? DATASET_TYPE_UNSTRUCTURED;
+        if ($dataset->type == DatasetManager::DATASET_TYPE_UNSTRUCTURED) {
+            ubiq_debug($creds, 'Key caching configuration for ' . DatasetManager::DATASET_TYPE_UNSTRUCTURED . ' is ' . $creds::$config['key_caching']['unstructured']);
 
-        if ($type == DATASET_TYPE_UNSTRUCTURED) {
-            ubiq_debug($creds, 'Key caching configuration for ' . DATASET_TYPE_UNSTRUCTURED . ' is ' . $creds->config['key_caching']['unstructured']);
+            return ($creds::$config['key_caching']['unstructured']);
+        }
 
-            return ($creds->config['key_caching']['unstructured']);
+        if ($dataset->type == DatasetManager::DATASET_TYPE_STRUCTURED) {
+            ubiq_debug($creds, 'Key caching configuration for ' . DatasetManager::DATASET_TYPE_STRUCTURED . ' is ' . $creds::$config['key_caching']['structured']);
+
+            return ($creds::$config['key_caching']['structured']);
         }
 
         return true;
@@ -200,7 +227,7 @@ class KeyManager
      *
      * @param Credentials $creds    Credentials object to operate on
      * @param Dataset     $dataset  Dataset to get a key for
-     * @param bool        $no_cache Force no-cache
+     * @param bool        $no_cache Force no-cache, used in "multiple_uses" context
      * 
      * @return Array of key data
      */
@@ -222,38 +249,72 @@ class KeyManager
         $cache = $this->getKeyDefault($creds, $dataset);
 
         if (empty($cache)) {
-            ubiq_debug($creds, 'Getting encryption key from backend for ' . $dataset->name);
+            ubiq_debug($creds, 'Getting encryption key from backend for ' . $dataset->type . ' ' . $dataset->name);
 
             $http = new Request(
                 $creds->getPapi(), $creds->getSapi()
             );
 
-            $resp = $http->post(
-                $creds->getHost() . '/api/v0/encryption/key',
-                json_encode(['uses' => 1]),
-                'application/json'
-            );
-
-            if (!$resp['success']) {
-                throw new \Exception(
-                    'Request for encryption key returned ' . $resp['status']
+            if ($dataset->type == DatasetManager::DATASET_TYPE_UNSTRUCTURED) {
+                $resp = $http->post(
+                    $creds->getHost() . '/api/v0/encryption/key',
+                    json_encode(['uses' => 1]),
+                    'application/json'
                 );
 
-                return;
+                if (!$resp['success']) {
+                    throw new \Exception(
+                        'Request for ' . $dataset->type . ' encryption key returned ' . $resp['status']
+                    );
+    
+                    return;
+                }
+                
+                $json = json_decode($resp['content'], true);
+    
+                $cache = [
+                    'key_idx'       => md5($json['encrypted_data_key']),
+                    '_key_enc'      => base64_decode($json['encrypted_data_key']),
+                    '_key_enc_prv'  => $json['encrypted_private_key'],
+                    '_key_raw'      => base64_decode($json['wrapped_data_key']),
+                    '_session'      => $json['encryption_session'],
+                    '_fingerprint'  => $json['key_fingerprint'],
+                    '_algorithm'    => new Algorithm(strtolower($json['security_model']['algorithm'])),
+                    '_fragment'     => $json['security_model']['enable_data_fragmentation']
+                ];
             }
-            
-            $json = json_decode($resp['content'], true);
+            elseif ($dataset->type == DatasetManager::DATASET_TYPE_STRUCTURED) {
+                $resp = $http->get(
+                    $creds->getHost() . '/api/v0/fpe/key?papi=' . urlencode($creds->getPapi()) . '&ffs_name=' . urlencode($dataset->name),
+                    'application/json'
+                );
 
-            $cache = [
-                'key_idx'       => md5($json['encrypted_data_key']),
-                '_key_enc'      => base64_decode($json['encrypted_data_key']),
-                '_key_enc_prv'  => $json['encrypted_private_key'],
-                '_key_raw'      => base64_decode($json['wrapped_data_key']),
-                '_session'      => $json['encryption_session'],
-                '_fingerprint'  => $json['key_fingerprint'],
-                '_algorithm'    => new Algorithm(strtolower($json['security_model']['algorithm'])),
-                '_fragment'     => $json['security_model']['enable_data_fragmentation']
-            ];
+                if (!$resp['success']) {
+                    throw new \Exception(
+                        'Request for ' . $dataset->type . ' encryption key returned ' . $resp['status']
+                    );
+    
+                    return;
+                }
+                
+                $json = json_decode($resp['content'], true);
+
+                // the idx is md5-d and encoded for consistency with the unstructured keys and not making changes
+                // in how the keycache looks up and stores them; it makes the index of the key cache a bit more
+                // obfuscated than just a key number when debugging, but is simpler code
+                // same goes for why we put key_number in _key_enc
+
+                $cache = [
+                    'key_idx'       => md5(base64_encode($json['key_number'])),
+                    '_key_enc'      => $json['key_number'],
+                    '_key_enc_prv'  => $json['encrypted_private_key'],
+                    '_key_raw'      => base64_decode($json['wrapped_data_key']),
+                    '_session'      => NULL,
+                    '_fingerprint'  => NULL,
+                    '_algorithm'    => new Algorithm('ff1'),
+                    '_fragment'     => NULL,
+                ];
+            }
 
             ubiq_debug($creds, 'Got encryption key from backend for ' . $dataset->name);
             
@@ -262,13 +323,114 @@ class KeyManager
 
             // if we're caching an encryption key, also cache it as the default
             if ($this->_shouldCache($creds, $dataset) && !$no_cache) {
-                $creds::$cachemanager->copy(CacheManager::CACHE_TYPE_KEYS, $dataset->name . '-keys-' . $cache['key_idx'], $dataset->name . '-keys-default');
+                ubiq_debug($creds, 'Caching as default encryption key for ' . $dataset->name . ' ' . $cache['key_idx']);
+                ubiq_debug($creds, 'Setting TTL for default encryption key for ' . $dataset->name . ' ' . $cache['key_idx'] . ' to ' . self::getCacheTTL($creds) . ' seconds');
+
+                $creds::$cachemanager::copy(
+                    CacheManager::CACHE_TYPE_KEYS,
+                    $dataset->name . '-keys-' . $cache['key_idx'],
+                    $dataset->name . '-keys-default'
+                );
+                
+                $creds::$cachemanager::setTTL(
+                    CacheManager::CACHE_TYPE_KEYS,
+                    $dataset->name . '-keys-default',
+                    strtotime('+' . self::getCacheTTL($creds) . ' seconds')
+                );
             }
         }
 
         ubiq_debug($creds, 'Finished getEncryptionKey for ' . $dataset->name);
 
         return $cache;
+    }
+
+    /**
+     * Get all encryption keys for a structured dataset for use with encryptForSearch
+     * Will cache if appropriate
+     *
+     * @param Credentials   $creds          Credentials object to operate on
+     * @param var           $dataset_names  String or array of strings for datasets to get a key for
+     * 
+     * @return array Array of keys data
+     */
+    public function getAllEncryptionKeys(
+        Credentials $creds = null,
+        $dataset_names = null
+    ) {
+        if (!is_array($dataset_names)) {
+            $dataset_names = [$dataset_names];
+        }
+
+        ubiq_debug($creds, 'Starting getAllEncryptionKeys for ' . implode(',', $dataset_names));
+
+        if (!$creds) {
+            throw new \Exception(
+                'No valid credentials'
+            );
+            
+            return;
+        }
+
+        $http = new Request(
+            $creds->getPapi(), $creds->getSapi()
+        );
+        
+        $resp = $http->get(
+            $creds->getHost() . '/api/v0/fpe/def_keys?papi=' . urlencode($creds->getPapi()) . '&ffs_name=' . urlencode(implode(',', $dataset_names)),
+            'application/json'
+        );
+
+        if (!$resp['success']) {
+            throw new \Exception(
+                'Request for ' . implode(',', $dataset_names) . ' encryption keys returned ' . $resp['status']
+            );
+
+            return;
+        }
+            
+        $json = json_decode($resp['content'], true);
+        $keys = [];
+        
+        foreach ($json as $dataset_name => $values) {
+            $dataset = new Dataset($dataset_name, null, DatasetManager::DATASET_TYPE_STRUCTURED, $values['ffs']);
+            
+            $creds::$cachemanager::set(
+                CacheManager::CACHE_TYPE_DATASET_CONFIGS, $dataset->name, $dataset
+            );
+
+            foreach ($values['keys'] as $idx => $key) {
+                $key_idx = md5(base64_encode($idx));
+
+                ubiq_debug($creds, 'Got encryption key ' . $idx . ' from backend for ' . $dataset_name . ' and caching to ' . $key_idx);
+
+                // same comments apply here as in getEncryptionKey
+                // key array is only the key material of the wrapped data key
+                $cache = [
+                    'key_idx'       => $key_idx,
+                    '_key_enc'      => $idx,
+                    '_key_enc_prv'  => $values['encrypted_private_key'],
+                    '_key_raw'      => base64_decode($key),
+                    '_session'      => NULL,
+                    '_fingerprint'  => NULL,
+                    '_algorithm'    => new Algorithm('ff1'),
+                    '_fragment'     => NULL,
+                ];
+
+                $cache = $this->cacheKey($creds, $dataset, $cache);
+
+                $keys[] = [
+                    'dataset' => $dataset,
+                    'key' => $cache
+                ];
+            }
+
+            ubiq_debug($creds, 'Got encryption keys from backend for ' . implode(',', $dataset_names));
+            
+            ubiq_debug($creds, 'Finished getAllEncryptionKeys for ' . implode(',', $dataset_names));
+
+            return $keys;
+        }
     }
 
     /**
@@ -295,9 +457,14 @@ class KeyManager
             
             return;
         }
-        $cache = $this->getKey($creds, $dataset, $key_headers['key_enc']);
+        
+        if ($dataset->type == DatasetManager::DATASET_TYPE_UNSTRUCTURED) {
+            $cache = $this->getKey($creds, $dataset, $key_headers['key_enc']);
 
-        if (empty($cache)) {
+            if (!empty($cache)) {
+                return $cache;
+            }
+
             ubiq_debug($creds, 'Getting decryption key from backend for ' . $dataset->name);
 
             $http = new Request(
@@ -314,14 +481,14 @@ class KeyManager
 
             if (!$resp) {
                 throw new \Exception(
-                    'Request for decryption key returned ' . $resp['status']
+                    'Request for  ' . $dataset->type . ' decryption key returned ' . $resp['status']
                 );
             }
 
             $json = json_decode($resp['content'], true);
 
             $cache = [
-                'key_idx'       => base64_encode(md5($key_headers['key_enc'])),
+                'key_idx'       => md5($key_headers['key_enc']),
                 '_key_enc'      => $key_headers['key_enc'],
                 '_key_enc_prv'  => $json['encrypted_private_key'] ?? NULL,
                 '_key_raw'      => base64_decode($json['wrapped_data_key'] ?? NULL),
@@ -329,12 +496,49 @@ class KeyManager
                 '_fingerprint'  => $json['key_fingerprint'] ?? NULL,
                 '_algorithm'    => new Algorithm($key_headers['algoid']),
             ];
-
-            ubiq_debug($creds, 'Got decryption key from backend for ' . $dataset->name);
-
-            // the return does decryption of key as needed
-            $cache = $this->cacheKey($creds, $dataset, $cache);
         }
+        elseif ($dataset->type == DatasetManager::DATASET_TYPE_STRUCTURED) {
+            $cache = $this->getKey($creds, $dataset, $key_headers['key_number']);
+
+            if (!empty($cache)) {
+                return $cache;
+            }
+
+            ubiq_debug($creds, 'Getting decryption key from backend for ' . $dataset->name);
+
+            $http = new Request(
+                $creds->getPapi(), $creds->getSapi()
+            );
+
+            $resp = $http->get(
+                $creds->getHost() . '/api/v0/fpe/key?papi=' . urlencode($creds->getPapi()) . '&ffs_name=' . urlencode($dataset->name) . '&key_number=' . $key_headers['key_number'],
+                'application/json'
+            );
+
+            if (!$resp) {
+                throw new \Exception(
+                    'Request for  ' . $dataset->type . ' decryption key returned ' . $resp['status']
+                );
+            }
+
+            $json = json_decode($resp['content'], true);
+
+            $cache = [
+                'key_idx'       => md5(base64_encode($json['key_number'])),
+                '_key_enc'      => NULL,
+                '_key_enc_prv'  => $json['encrypted_private_key'],
+                '_key_raw'      => base64_decode($json['wrapped_data_key']),
+                '_session'      => NULL,
+                '_fingerprint'  => NULL,
+                '_algorithm'    => new Algorithm('ff1'),
+                '_fragment'     => NULL,
+            ];
+        }
+
+        ubiq_debug($creds, 'Got decryption key from backend for ' . $dataset->name);
+
+        // the return does decryption of key as needed
+        $cache = $this->cacheKey($creds, $dataset, $cache);
 
         ubiq_debug($creds, 'Finished getDecryptionKey for ' . $dataset->name);
 
